requirements.txt
astroid==1.5.3
colorama==0.3.9
Django==1.11.5
django-bootstrap-form==3.3
django-bootstrap3==9.0.0
django-favicon==0.1.3
django-gravatar2==1.4.2
django-storages==1.6.5
djangorestframework==3.6.4
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
Pillow==4.2.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.2
six==1.11.0
wrapt==1.10.11
